[feat] #1 프로젝트 - 프로젝트 생성 시 admin이 아니면 예외 처리

+++ src/main/java/com/example/IssueTrackingSystem/service/ProjectService/ProjectCommandServiceImpl.java
@@ -1,5 +1,8 @@
 package com.example.IssueTrackingSystem.service.ProjectService;
 
+import com.example.IssueTrackingSystem.apiPayload.code.status.ErrorStatus;
+import com.example.IssueTrackingSystem.apiPayload.exception.handler.IssueHandler;
+import com.example.IssueTrackingSystem.apiPayload.exception.handler.ProjectHandler;
 import com.example.IssueTrackingSystem.converter.ProjectConverter;
 import com.example.IssueTrackingSystem.domain.entity.Project;
 import com.example.IssueTrackingSystem.domain.entity.User;
@@ -10,6 +13,9 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+
+import static com.example.IssueTrackingSystem.domain.enums.UserRole.ADMIN;
+import static com.example.IssueTrackingSystem.domain.enums.UserRole.TESTER;
 
 @Service
 @RequiredArgsConstructor
@@ -25,6 +31,11 @@ public class ProjectCommandServiceImpl implements ProjectCommandService{
         User getUser = userRepository.findById(userId).get();
         newProject.setUser(getUser);
 
+        // admin만 프로젝트 생성 가능
+        if(getUser.getUserRole() != ADMIN){
+            throw new ProjectHandler(ErrorStatus.PROJECT_CREATE_UNAUTHORIZED);
+        }
+
         // Project entity db에 저장
         Project savedProject = projectRepository.save(newProject);
 
@@ -32,7 +43,13 @@ public class ProjectCommandServiceImpl implements ProjectCommandService{
     }
 
     @Override
-    public Project updateProject(Long projectId, ProjectRequestDTO.UpdateProjectDTO request) {
+    public Project updateProject(Long userId, Long projectId, ProjectRequestDTO.UpdateProjectDTO request) {
+        // admin만 프로젝트 수정 가능
+        User getUser = userRepository.findById(userId).get();
+        if(getUser.getUserRole() != ADMIN){
+            throw new ProjectHandler(ErrorStatus.PROJECT_UPDATE_UNAUTHORIZED);
+        }
+
         Project updateProject = projectRepository.findById(projectId).get();
         updateProject.update(request);
 
@@ -40,7 +57,13 @@ public class ProjectCommandServiceImpl implements ProjectCommandService{
     }
 
     @Override
-    public void deleteProject(Long projectId) {
+    public void deleteProject(Long userId, Long projectId) {
+        // admin만 프로젝트 삭제 가능
+        User getUser = userRepository.findById(userId).get();
+        if(getUser.getUserRole() != ADMIN){
+            throw new ProjectHandler(ErrorStatus.PROJECT_DELETE_UNAUTHORIZED);
+        }
+
         Project deleteProject = projectRepository.findById(projectId).get();
         projectRepository.delete(deleteProject);
     }
